renamed "LabyrinthEngine" -> "labyrinth_engine"
renamed "LabyrinthObjects" -> "labyrinth_objects"
