--- LabyrinthEngine/__init__.py
@@ -1,4 +1,0 @@
-from LabyrinthEngine.load_save import load_lrmap, load_lrsave
-from LabyrinthEngine.labyrinth import Labyrinth
-from LabyrinthEngine.labyrinth_object import LabyrinthObject, get_attr_safe
-from LabyrinthEngine.lr_types import Location, Item, Player, Creature
