--- LabyrinthObjects/Vanilla/__init__.py
@@ -1,13 +1,0 @@
-from LabyrinthObjects.Vanilla.ammo import Ammo
-from LabyrinthObjects.Vanilla.arsenal import Arsenal
-from LabyrinthObjects.Vanilla.bear import Bear
-from LabyrinthObjects.Vanilla.bomb import Bomb
-from LabyrinthObjects.Vanilla.empty import EmptyLocation
-from LabyrinthObjects.Vanilla.exit import Exit
-from LabyrinthObjects.Vanilla.first_aid_post import FirstAidPost
-from LabyrinthObjects.Vanilla.gun import Gun
-from LabyrinthObjects.Vanilla.health import Health
-from LabyrinthObjects.Vanilla.hole import Hole
-from LabyrinthObjects.Vanilla.legs import Legs
-from LabyrinthObjects.Vanilla.treasure import Treasure
-from LabyrinthObjects.Vanilla.walls import Wall, Outside
