+++ labyrinth_engine/common_functions.py
