+++ labyrinth_objects/DevelopersBacklog/see_and_foresee.py
