+++ labyrinth_objects/Vanilla/__init__.py
@@ -1,0 +1,13 @@
+from labyrinth_objects.Vanilla.ammo import Ammo
+from labyrinth_objects.Vanilla.arsenal import Arsenal
+from labyrinth_objects.Vanilla.bear import Bear
+from labyrinth_objects.Vanilla.bomb import Bomb
+from labyrinth_objects.Vanilla.empty import EmptyLocation
+from labyrinth_objects.Vanilla.exit import Exit
+from labyrinth_objects.Vanilla.first_aid_post import FirstAidPost
+from labyrinth_objects.Vanilla.gun import Gun
+from labyrinth_objects.Vanilla.health import Health
+from labyrinth_objects.Vanilla.hole import Hole
+from labyrinth_objects.Vanilla.legs import Legs
+from labyrinth_objects.Vanilla.treasure import Treasure
+from labyrinth_objects.Vanilla.walls import Wall, Outside
