+++ labyrinth_objects/Vanilla/ammo.py
@@ -1,4 +1,4 @@
-from LabyrinthEngine import Item
+from labyrinth_engine import Item
 
 
 class Ammo(Item):
