+++ labyrinth_objects/Vanilla/default_settings.json
