+++ labyrinth_objects/Vanilla/dock.md
