+++ labyrinth_objects/Vanilla/first_aid_post.py
@@ -1,4 +1,4 @@
-from LabyrinthEngine import Location
+from labyrinth_engine import Location
 
 
 class FirstAidPost(Location):
