+++ tmp/Примерчик.map.json
@@ -1,0 +1,234 @@
+{
+    "locations":
+    [
+        {
+            "module":"LabyrinthObjects.Vanilla",
+            "class_name":"Outside"
+        },
+        {
+            "module":"LabyrinthObjects.Vanilla",
+            "class_name":"EmptyLocation"
+        },
+        {
+            "module":"LabyrinthObjects.Vanilla",
+            "class_name":"Hole"
+        },
+        {
+            "module":"LabyrinthObjects.Vanilla",
+            "class_name":"EmptyLocation"
+        },
+        {
+            "module":"LabyrinthObjects.Vanilla",
+            "class_name":"Arsenal"
+        },
+        {
+            "module":"LabyrinthObjects.Vanilla",
+            "class_name":"EmptyLocation"
+        },
+        {
+            "module":"LabyrinthObjects.Vanilla",
+            "class_name":"EmptyLocation"
+        },
+        {
+            "module":"LabyrinthObjects.Vanilla",
+            "class_name":"Hole"
+        },
+        {
+            "module":"LabyrinthObjects.Vanilla",
+            "class_name":"FirstAidPost"
+        },
+        {
+            "module":"LabyrinthObjects.Vanilla",
+            "class_name":"Exit"
+        },
+        {
+            "module":"LabyrinthObjects.Vanilla",
+            "class_name":"Wall"
+        },
+        {
+            "module":"LabyrinthObjects.Vanilla",
+            "class_name":"Wall"
+        }
+    ],
+    "items":
+    [
+        {
+            "module":"LabyrinthObjects.Vanilla",
+            "class_name":"Legs"
+        },
+        {
+            "module":"LabyrinthObjects.Vanilla",
+            "class_name":"Gun"
+        },
+        {
+            "module":"LabyrinthObjects.Vanilla",
+            "class_name":"Bomb"
+        },
+        {
+            "module":"LabyrinthObjects.Vanilla",
+            "class_name":"Treasure"
+        },
+        {
+            "module":"LabyrinthObjects.Vanilla",
+            "class_name":"Health"
+        },
+        {
+            "module":"LabyrinthObjects.Vanilla",
+            "class_name":"Ammo"
+        }
+    ],
+    "creatures":
+    [
+        {
+            "module":"LabyrinthObjects.Vanilla",
+            "class_name":"Bear"
+        }
+    ],
+    "adjacence_list":
+    [
+        {},
+        {
+            "up":0,
+            "down":5,
+            "right":10,
+            "left":0
+        },
+        {
+            "up":0,
+            "down":6,
+            "right":3,
+            "left":10
+        },
+        {
+            "up":0,
+            "down":7,
+            "right":4,
+            "left":2
+        },
+        {
+            "up":0,
+            "down":8,
+            "right":0,
+            "left":3
+        },
+        {
+            "up":1,
+            "down":0,
+            "right":6,
+            "left":0
+        },
+        {
+            "up":2,
+            "down":9,
+            "right":11,
+            "left":5
+        },
+        {
+            "up":3,
+            "down":0,
+            "right":8,
+            "left":11
+        },
+        {
+            "up":4,
+            "down":0,
+            "right":0,
+            "left":7
+        },
+        {
+            "up":6,
+            "down":0,
+            "right":0,
+            "left":0
+        },
+        {
+            "right":2,
+            "left":1
+        }, 
+        {
+            "right":7,
+            "left":6
+        }
+    ],
+    "settings":
+    {
+        "locations": [
+            {
+                "name":"0"
+            }, 
+            {
+                "name":"1", 
+                "enter_msg": {"ru":"Вы находитесь в тёмной комнате. Тут холодно и сыро."}
+            }, 
+            {
+                "name": "2",
+                "fall_to":7
+            },
+            {
+                "name":"3",
+                "enter_msg": {"ru":"В комнате лежит скелет. Вы его уже видели.\nСкелет по прежнему выглядит угрожающе."},
+                "first_enter_msg": {"ru":"Вы находитесь в очередной комнате.\nЗдесь вас пробирает дрожь. Скелет у стенки выглядит угрожающе."}
+            }, 
+            {
+                "name":"4"
+            }, 
+            {
+                "name":"5"
+            }, 
+            {
+                "name":"6"
+            }, 
+            {
+                "name":"7",
+                "fall_to":2
+            }, 
+            {
+                "name":"8"
+            }, 
+            {
+                "name":"9"
+            }, 
+            {
+                "name":"10"
+            }, 
+            {
+                "name":"11"
+            }],
+        "items": [
+            {
+                "name":"0"
+            }, 
+            {
+                "name":"1"
+            }, 
+            {
+                "name":"2"
+            },
+            {
+                "name":"3",
+                "is_true": true,
+                "position":3
+            }, 
+            {
+                "name": "4"
+            }, 
+            {
+                "name": "5"
+            }
+        ],
+        "creatures": [
+            {
+                "name":"Медведь",
+                "position":4
+            }
+        ],
+        "player": {
+            "flags":["drop_items_when_injured"],
+            "start_positions":
+            {
+                "distribution": "random",
+                "from": [1, 3, 5]
+            }
+        }
+    }
+}
